requirements.txt
pytorch==2.1.2
pytorch-lightning==2.1.3